Move display application link test to integration test
